@@ -12,15 +12,20 @@ repository = "https://github.com/jcarrete5/othello-ai"
 keywords = ["othello", "ai"]
 
 [tool.poetry.dependencies]
-python = "^3.7"
-torch = "^1.3.1"
+python = "^3.9"
 
 [tool.poetry.dev-dependencies]
-pylint = "^2.4"
 pytest = "^5.2.1"
+flake8 = "^3.9.0"
+isort = "^5.8.0"
+black = "^20.8b1"
 
 [tool.poetry.scripts]
 othello = "othelloai.main:start_ui"
+
+[tool.isort]
+profile = "black"
+multi_line_output = 3
 
 [build-system]
 requires = ["poetry>=0.12"]
