@@ -1,5 +1,8 @@
 import logging
 import os
-logging.basicConfig(style='{',
-                    format='[{name}:{levelname}] - {threadName} - In {funcName}: {message}',
-                    level=getattr(logging, os.getenv('LOG', ''), logging.WARNING))
+
+logging.basicConfig(
+    style="{",
+    format="[{name}:{levelname}] - {threadName} - In {funcName}: {message}",
+    level=getattr(logging, os.getenv("LOG", ""), logging.WARNING),
+)
