@@ -1,7 +1,6 @@
-"""
-Strategies and implementation details for the AI.
-"""
+"""Strategies and implementation details for the AI."""
 from __future__ import annotations
+
 from random import choice as chooseFrom
 from typing import TYPE_CHECKING
 
@@ -9,8 +8,9 @@ if TYPE_CHECKING:
     from . import bitboard as bb
     from .game import BoardState
 
+
 async def random(state: BoardState) -> bb.Position:
-    """ Returns a random valid move or None if no moves are possible. """
+    """Return a random valid move or None if no moves are possible."""
     try:
         return chooseFrom(state.valid_moves())
     except IndexError:
