@@ -1,13 +1,14 @@
 import logging
+
 from othelloai import ui
 
-_LOGGER = logging.getLogger('othello')
+_LOGGER = logging.getLogger("othello")
 
 
 def start_ui():
     try:
         ui.loop()  # Blocked until application is closed or an error occurs
     except KeyboardInterrupt:
-        _LOGGER.info('Quitting application via KeyboardInterrupt...')
+        _LOGGER.info("Quitting application via KeyboardInterrupt...")
     finally:
         logging.shutdown()
