@@ -1,23 +1,19 @@
 from __future__ import annotations
-from typing import TYPE_CHECKING, Callable
+
 import asyncio
 import enum
+from typing import TYPE_CHECKING, Callable
 
 if TYPE_CHECKING:
-    from .game import BoardState
     from .bitboard import Position
+    from .game import BoardState
 
-
-class Color(enum.Enum):
-    """
-    Represents color of pieces. There can only be two colors.
-    """
-    BLACK = enum.auto()
-    WHITE = enum.auto()
+Color = enum.Enum("Color", "BLACK WHITE")
 
 
 class Player:
-    """ Interface for playing and retrieving moves asynchronously. """
+    """Interface for playing and retrieving moves asynchronously."""
+
     def __init__(self, color: Color):
         self._color = color
         self._loop = asyncio.get_running_loop()
@@ -38,8 +34,11 @@ class Player:
     def move(self, value: Position):
         def callback():
             if self._move.done():
-                raise RuntimeError(f'Attempt to set {self} move but it has already been set')
+                raise RuntimeError(
+                    f"Attempt to set {self} move but it has already been set"
+                )
             self._move.set_result(value)
+
         self._loop.call_soon_threadsafe(callback)
 
     @move.deleter
@@ -47,15 +46,17 @@ class Player:
         self._move = self._loop.create_future()
 
     def __repr__(self):
-        return f'Player({self._color})'
+        return f"Player({self._color})"
 
 
 class AIPlayer(Player):
-    def __init__(self, color: Color, state: BoardState, strat: Callable[[BoardState], Position]):
+    def __init__(
+        self, color: Color, state: BoardState, strat: Callable[[BoardState], Position]
+    ):
         super().__init__(color)
         self._strat = strat
         self._state = state
 
-    @Player.move.getter  # pylint: disable=no-member
-    async def move(self):  # pylint: disable=invalid-overridden-method
+    @Player.move.getter
+    async def move(self):
         return await self._strat(self._state)
